# Wasm runtime: make resuming continuations more efficient