… runtime/wasm/effect.wat