This was done in #1765 for the JavaScript runtime.
See corresponding OCaml PR ocaml/ocaml#12735

fix https://github.com/ocsigen/js_of_ocaml/issues/1779