@hhugo 2025-03-18
I used the deep_state benchmarks and show the following improvement

compile the bench with:
```
WASM_OF_OCAML=true dune build  compiler/tests-jsoo/lib-effects/deep_state.bc.wasm.js --profile with-effects
```

before:
```
$ time node _build/default/compiler/tests-jsoo/lib-effects/deep_state.bc.wasm.js 128 500_000
Running deepstate: depth=128 ops=500000

real	0m6,003s
user	0m5,980s
sys	0m0,033s
```
after:
```
$ time node _build/default/compiler/tests-jsoo/lib-effects/deep_state.bc.wasm.js 128 500_000
Running deepstate: depth=128 ops=500000

real	0m2,359s
user	0m2,342s
sys	0m0,025s
```

stack switching is however very slow
```
$ time node --experimental-wasm-stack-switching _build/default/compiler/tests-jsoo/lib-effects/deep_state.bc.wasm.js 128 500_000
Running deepstate: depth=128 ops=500000

real	1m11,467s
user	1m11,388s
sys	0m0,122s
```