@vouillon 2025-03-18
> stack switching is however very slow

Yes, that's expected. It goes through the JavaScript event loop for each stack switch.