@vouillon 2025-03-21
> Should the doc/readme say something about this ?

We have this at the moment:
https://github.com/ocsigen/js_of_ocaml/blob/55c622862c506a2d403d3f1d5d5c6bfda6bf45d7/manual/wasm_overview.wiki?plain=1#L56