@vouillon 2025-04-01
You can have a look at #1765 and #1891.

We have some changes to make in perform/reperform since we were storing the fibers in reversed order in the continuation. They were [already stored in the right order in the OCaml runtime](https://github.com/ocaml/ocaml/commit/fc6cacea67bd965bf3765975e7b860d551d824b9). Also, we need to [store the tail of the stack in the continuation](https://github.com/ocaml/ocaml/pull/12309/files#diff-463184d7ba0c7e61cae25a864aaf737689642847d21a94dbbf4c1abe6cfdbc54).

Separating perform and reperform also allows some simplifications regarding unhandled effects: for perform, we can just raise the exception right away, and for reperform, we can resume the continuation with an exception. One no longer needs to put a default handler at the top of the current stack.