@vouillon 2025-04-02
> Do we have an effects benchmark to assess the impact of this PR?

@hhugo [did some measurements](https://github.com/ocsigen/js_of_ocaml/pull/1892#issuecomment-2732853198).